requirements.txt
streamlit>=1.0
qrcode>=7.3
Pillow>=9.0
